Claude_Code_Agentic_AI/requirements.txt
kubernetes>=28.0.0
pyyaml>=6.0
python-dateutil>=2.8.2
ollama>=0.1.0